Fix FileNotFoundException handling
